Se agrega la funcionalidad de editar noticia

+++ frontend/simple-news/src/views/HomeView.vue
@@ -4,8 +4,8 @@
     <ul>
       <li v-for="news in newsList" :key="news.id">
         <strong>{{ news.title }}</strong>
-        
-        <button @click="handleDelete(news.id)">❌ Eliminar</button>
+        <button v-if="isLoggedIn" @click="goToEdit(news.id)">📝 Editar</button>
+        <button v-if="isLoggedIn" @click="handleDelete(news.id)">❌ Eliminar</button>
         <p>
           {{ news.content.length > 200 ? news.content.slice(0, 200) + '...' : news.content }}
           <template v-if="news.content.length > 200">
@@ -29,6 +29,9 @@
 <script setup lang="ts">
 import { ref, onMounted, watch } from 'vue'
 import { fetchNews, deleteNews } from '../services/newsService'
+import { useAuthStore } from '../stores/authStore'
+import { useRouter } from 'vue-router'
+import { computed } from 'vue'
 import type { News } from '../models/newsModel'
 
 const newsList = ref<News[]>([])
@@ -37,6 +40,8 @@ const error = ref('')
 const page = ref(1)
 const limit = 10
 const totalPages = ref(1)
+const auth = useAuthStore()
+const isLoggedIn = computed(() => !!auth.token)
 
 async function loadNews() {
   loading.value = true
@@ -57,7 +62,17 @@ function changePage(newPage: number) {
   }
 }
 
+const router = useRouter()
+
+function goToEdit(newsId: number) {
+  router.push(`/dashboard/${newsId}`)
+}
+
 async function handleDelete(id: number) {
+  if (!auth.token) {
+    alert('Debes iniciar sesión para eliminar noticias.')
+    return
+  }
   if (confirm('¿Seguro que quieres eliminar esta noticia?')) {
     try {
       await deleteNews(id)
